+++ lib/mauth/client.rb
@@ -7,6 +7,7 @@ require 'mauth/core_ext'
 require 'mauth/autoload'
 require 'mauth/dice_bag/mauth_templates'
 require 'mauth/version'
+require 'mauth/v2'
 
 module MAuth
   class Client
@@ -303,16 +304,6 @@ module MAuth
         { 'X-MWS-Authentication' => "#{MWS_TOKEN} #{client_app_uuid}:#{signature}", 'X-MWS-Time' => attributes[:time] }
       end
 
-      def signed_headers_v2(object, attributes = {})
-        attributes = { time: Time.now.to_i.to_s, app_uuid: client_app_uuid }.merge(attributes)
-        string_to_sign = object.string_to_sign_v2(attributes)
-        signature = self.signature(string_to_sign)
-        {
-          'MCC-Authentication' => "#{MWSV2_TOKEN} #{client_app_uuid}:#{signature}#{AUTH_HEADER_DELIMITER}",
-          'MCC-Time' => attributes[:time]
-        }
-      end
-
       def signature(string_to_sign)
         assert_private_key(UnableToSignError.new('mAuth client cannot sign without a private key!'))
         Base64.encode64(private_key.private_encrypt(string_to_sign)).delete("\n")
@@ -417,41 +408,6 @@ module MAuth
         raise InauthenticError, msg
       end
 
-      # V2 helpers
-      def authenticate_v2!(object)
-        time_valid_v2!(object)
-        token_valid_v2!(object)
-        signature_valid_v2!(object)
-      end
-
-      def authentication_present_v2(object)
-        !object.mcc_authentication.nil? || object.mcc_authentication&.match?(/\S/)
-      end
-
-      def authentication_present_v2!(object)
-        return if authentication_present_v2(object)
-
-        msg = 'Authentication Failed. No mAuth signature present; MCC-Authentication header is blank.'
-        log_mauth_not_present(object, msg)
-        raise MauthNotPresent, msg
-      end
-
-      def time_valid_v2!(object)
-        if object.mcc_time.nil?
-          msg = 'Time verification failed. No MCC-Time present.'
-          log_inauthentic(object, msg)
-          raise InauthenticError, msg
-        end
-        time_within_valid_range!(object, object.mcc_time.to_i)
-      end
-
-      def token_valid_v2!(object)
-        return if object.signature_token == MWSV2_TOKEN
-
-        msg = "Token verification failed. Expected #{MWSV2_TOKEN}; token was #{object.signature_token}"
-        log_inauthentic(object, msg)
-        raise InauthenticError, msg
-      end
     end
     include Authenticator
 
@@ -483,48 +439,6 @@ module MAuth
 
         # reset the object original request_uri, just in case we need it again
         object.attributes_for_signing[:request_url] = original_request_uri
-
-        pubkey = OpenSSL::PKey::RSA.new(retrieve_public_key(object.signature_app_uuid))
-        begin
-          actual = pubkey.public_decrypt(Base64.decode64(object.signature))
-        rescue OpenSSL::PKey::PKeyError
-          raise InauthenticError, "Public key decryption of signature failed!\n#{$!.class}: #{$!.message}"
-        end
-
-        unless expected_no_reencoding == actual || expected_euresource_style_reencoding == actual || expected_for_percent_reencoding == actual
-          raise InauthenticError, "Signature verification failed for #{object.class}"
-        end
-      end
-
-      def signature_valid_v2!(object)
-        # We are in an unfortunate situation in which Euresource is percent-encoding parts of paths, but not
-        # all of them.  In particular, Euresource is percent-encoding all special characters save for '/'.
-        # Also, unfortunately, Nginx unencodes URIs before sending them off to served applications, though
-        # other web servers (particularly those we typically use for local testing) do not.  The various forms
-        # of the expected string to sign are meant to cover the main cases.
-        # TODO:  Revisit and simplify this unfortunate situation.
-
-        original_request_uri = object.attributes_for_signing[:request_url]
-        original_query_string = object.attributes_for_signing[:query_string]
-
-        # craft an expected string-to-sign without doing any percent-encoding
-        expected_no_reencoding = object.string_to_sign_v2(time: object.mcc_time, app_uuid: object.signature_app_uuid)
-
-        # do a simple percent reencoding variant of the path
-        expected_for_percent_reencoding = object.string_to_sign_v2(
-          time: object.mcc_time,
-          app_uuid: object.signature_app_uuid,
-          request_url: CGI.escape(original_request_uri.to_s),
-          query_string: CGI.escape(original_query_string.to_s)
-        )
-
-        # do a moderately complex Euresource-style reencoding of the path
-        expected_euresource_style_reencoding = object.string_to_sign_v2(
-          time: object.mcc_time,
-          app_uuid: object.signature_app_uuid,
-          request_url: euresource_escape(original_request_uri.to_s),
-          query_string: euresource_escape(original_query_string.to_s)
-        )
 
         pubkey = OpenSSL::PKey::RSA.new(retrieve_public_key(object.signature_app_uuid))
         begin
@@ -636,22 +550,6 @@ module MAuth
           'request_url' => object.attributes_for_signing[:request_url],
           'request_time' => object.x_mws_time,
           'b64encoded_body' => Base64.encode64(object.attributes_for_signing[:body] || '')
-        }
-        make_mauth_request(authentication_ticket)
-      end
-
-      # TODO: update mAuth to be able verify authentication tickets w V2 (MCC-413109)
-      def signature_valid_v2!(object)
-        raise ArgumentError, "Remote Authenticator can only authenticate requests; received #{object.inspect}" unless object.is_a?(MAuth::Request)
-        authentication_ticket = {
-          'verb' => object.attributes_for_signing[:verb],
-          'app_uuid' => object.signature_app_uuid,
-          'client_signature' => object.signature,
-          'request_url' => object.attributes_for_signing[:request_url],
-          'request_time' => object.mcc_time,
-          'b64encoded_body' => Base64.encode64(object.attributes_for_signing[:body] || ''),
-          'query_string' => object.attributes_for_signing[:query_string],
-          'token' => object.signature_token
         }
         make_mauth_request(authentication_ticket)
       end
